feat: add marawa types
