+++ package.json
@@ -35,25 +35,25 @@
   "devDependencies": {
     "@appuniversum/ember-appuniversum": "^3.7.0",
     "@babel/core": "^7.26.0",
-    "@babel/eslint-parser": "^7.25.9",
+    "@babel/eslint-parser": "^7.26.5",
     "@babel/plugin-proposal-decorators": "^7.25.9",
     "@changesets/changelog-github": "^0.5.0",
     "@changesets/cli": "^2.26.2",
-    "@ember-data-types/adapter": "5.4.0-alpha.131",
-    "@ember-data-types/graph": "5.4.0-alpha.131",
-    "@ember-data-types/json-api": "5.4.0-alpha.131",
-    "@ember-data-types/legacy-compat": "5.4.0-alpha.131",
-    "@ember-data-types/model": "5.4.0-alpha.131",
-    "@ember-data-types/request": "5.4.0-alpha.131",
-    "@ember-data-types/request-utils": "5.4.0-alpha.131",
-    "@ember-data-types/serializer": "5.4.0-alpha.131",
-    "@ember-data-types/store": "5.4.0-alpha.131",
-    "@ember-data-types/tracking": "5.4.0-alpha.131",
-    "@ember-data/store": "^4.12.8",
+    "@ember-data-types/adapter": "5.4.0-alpha.143",
+    "@ember-data-types/graph": "5.4.0-alpha.143",
+    "@ember-data-types/json-api": "5.4.0-alpha.143",
+    "@ember-data-types/legacy-compat": "5.4.0-alpha.143",
+    "@ember-data-types/model": "5.4.0-alpha.143",
+    "@ember-data-types/request": "5.4.0-alpha.143",
+    "@ember-data-types/request-utils": "5.4.0-alpha.143",
+    "@ember-data-types/serializer": "5.4.0-alpha.143",
+    "@ember-data-types/store": "5.4.0-alpha.143",
+    "@ember-data-types/tracking": "5.4.0-alpha.143",
     "@ember/optional-features": "^2.1.0",
     "@ember/render-modifiers": "^2.1.0",
     "@ember/string": "^3.1.1",
     "@ember/test-helpers": "^3.3.1",
+    "@eslint/js": "^9.19.0",
     "@glimmer/component": "^1.1.2",
     "@glimmer/tracking": "^1.1.2",
     "@glint/core": "^1.5.0",
@@ -67,12 +67,12 @@
     "@lblod/ember-rdfa-editor-lblod-plugins": "28.0.0",
     "@lblod/template-uuid-instantiator": "^1.0.3",
     "@nullvoxpopuli/ember-composable-helpers": "^5.2.6",
+    "@rdfjs/types": "^2.0.1",
     "@release-it-plugins/lerna-changelog": "^6.0.0",
-    "@tsconfig/ember": "^3.0.8",
+    "@tsconfig/ember": "^3.0.9",
     "@types/eslint__js": "^8.42.3",
-    "@typescript-eslint/eslint-plugin": "^8.25.0",
-    "@typescript-eslint/parser": "^8.25.0",
-    "@warp-drive-types/core-types": "5.4.0-alpha.131",
+    "@types/uuid": "^10.0.0",
+    "@warp-drive-types/core-types": "5.4.0-alpha.143",
     "broccoli-asset-rev": "^3.0.0",
     "broccoli-plugin": "^4.0.7",
     "buffer": "^6.0.3",
@@ -99,10 +99,10 @@
     "ember-cli-terser": "^4.0.2",
     "ember-concurrency": "^4.0.2",
     "ember-config-helper": "^0.1.3",
-    "ember-data": "~4.12.0",
+    "ember-data": "~4.12.8",
     "ember-data-resources": "^5.0.0",
     "ember-data-table": "^2.1.0",
-    "ember-data-types": "5.4.0-alpha.131",
+    "ember-data-types": "5.4.0-alpha.143",
     "ember-feature-flags": "^6.0.0",
     "ember-fetch": "^8.1.2",
     "ember-intl": "^7.0.3",
@@ -121,11 +121,13 @@
     "ember-template-imports": "^4.1.2",
     "ember-template-lint": "^6.0.0",
     "ember-truth-helpers": "^4.0.3",
-    "eslint": "^9.17.0",
+    "eslint": "^9.19.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-ember": "^12.3.3",
+    "eslint-import-resolver-typescript": "^3.8.4",
+    "eslint-plugin-ember": "^12.5.0",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-n": "^17.15.1",
-    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-prettier": "^5.2.3",
     "eslint-plugin-qunit": "^8.1.2",
     "globals": "^16.0.0",
     "leaflet": "^1.9.4",
@@ -144,8 +146,8 @@
     "stylelint-config-standard": "^37.0.0",
     "stylelint-prettier": "^5.0.3",
     "tracked-built-ins": "^3.3.0",
-    "typescript": "^5.7.0",
-    "typescript-eslint": "^8.25.0",
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.23.0",
     "uuid": "^9.0.1",
     "webpack": "^5.95.0"
   },
