+++ tsconfig.json
@@ -6,7 +6,7 @@
   },
   "compilerOptions": {
     "allowJs": true,
-    "checkJs": false,
+    // "checkJs": true,
     // The combination of `baseUrl` with `paths` allows Ember's classic package
     // layout, which is not resolvable with the Node resolution algorithm, to
     // work with TypeScript.
@@ -18,18 +18,19 @@
     },
     "types": [
       "./node_modules/ember-source/types/stable",
-      "./node_modules/ember-data-types/unstable-preview-types",
-      "./node_modules/@ember-data-types/store/unstable-preview-types",
       "./node_modules/@ember-data-types/adapter/unstable-preview-types",
       "./node_modules/@ember-data-types/graph/unstable-preview-types",
       "./node_modules/@ember-data-types/json-api/unstable-preview-types",
       "./node_modules/@ember-data-types/legacy-compat/unstable-preview-types",
-      "./node_modules/@ember-data-types/request/unstable-preview-types",
-      "./node_modules/@ember-data-types/request-utils/unstable-preview-types",
       "./node_modules/@ember-data-types/model/unstable-preview-types",
+      "./node_modules/@ember-data-types/request-utils/unstable-preview-types",
+      "./node_modules/@ember-data-types/request/unstable-preview-types",
       "./node_modules/@ember-data-types/serializer/unstable-preview-types",
+      "./node_modules/@ember-data-types/store/unstable-preview-types",
       "./node_modules/@ember-data-types/tracking/unstable-preview-types",
-      "./node_modules/@warp-drive/core-types/unstable-preview-types"
+      "./node_modules/@warp-drive-types/core-types/unstable-preview-types",
+      "./node_modules/@warp-drive/core-types/unstable-preview-types",
+      "./node_modules/ember-data-types/unstable-preview-types"
     ]
   }
 }
