+++ types/@lblod/marawa/dom-helpers.d.ts
@@ -1,0 +1,2 @@
+export function findFirstNodeOfType(node: Node, type: string): Node | null;
+export function findAllNodesOfType(node: Node, type: string): Node[];
