+++ types/@lblod/marawa/ember-object-mock.d.ts
@@ -1,0 +1,1 @@
+export function warn(msg: string): void;
