+++ types/@lblod/marawa/triple.d.ts
@@ -1,0 +1,14 @@
+import type { Quad } from '@rdfjs/types/data-model';
+export default class Triple {
+  get subject(): string;
+  get predicate(): string;
+  get object(): string;
+  get datatype(): string;
+  get language(): string;
+
+  isEqual(other: Triple): boolean;
+
+  toNT(): string;
+
+  toQuad(): Quad;
+}
