+++ types/@lblod/template-uuid-instantiator.d.ts
