--- types/lblod/marawa/dom-helpers.d.ts
@@ -1,4 +1,0 @@
-declare module '@lblod/marawa/dom-helpers' {
-  export function findFirstNodeOfType(node: Node, type: string): Node | null;
-  export function findAllNodesOfType(node: Node, type: string): Node[];
-}
