--- types/lblod/marawa/ember-object-mock.d.ts
@@ -1,3 +1,0 @@
-declare module '@lblod/marawa/ember-object-mock' {
-  export function warn(msg: string): void;
-}
